pyproject.toml
[project]
name = "devops-analysis"
version = "0.1.0"
description = "Azure DevOps PR workflow with Supabase vector DB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.45",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.10",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "supabase>=2.18.1",
    "vecs[text-embedding]>=0.4.5",
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-java>=0.21.0",
    "tree-sitter-cpp>=0.22.0",
    "tree-sitter-c>=0.21.0",
    "tree-sitter-rust>=0.21.0",
    "tree-sitter-go>=0.21.0",
    "langchain-text-splitters>=0.3.11",
    "tree-sitter-swift>=0.0.1",
    "reportlab>=4.4.4",
    "markdown>=3.9",
    "weasyprint>=66.0",
    "xhtml2pdf>=0.2.17",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
